+++ src/models/alimentos.js
@@ -2,16 +2,20 @@
 import { DataTypes } from 'sequelize';
 import sequelize from '../config/database.js';
 
-const Alimento = sequelize.define('ALIMENTOS', {
-  id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  nombre: { type: DataTypes.STRING },
-  precio: { type: DataTypes.FLOAT },
-  descripcion: { type: DataTypes.STRING },
-  estado: { type: DataTypes.BOOLEAN, defaultValue: true }
-
-}, {
-  freezeTableName: true,
-  timestamps: false
-});
+const Alimento = sequelize.define(
+  "ALIMENTOS",
+  {
+    id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
+    nombre: { type: DataTypes.STRING },
+    // En Railway suele venir como DECIMAL / NUMERIC
+    precio: { type: DataTypes.DECIMAL },
+    descripcion: { type: DataTypes.STRING },
+    estado: { type: DataTypes.BOOLEAN, defaultValue: true }
+  },
+  {
+    freezeTableName: true,
+    timestamps: false
+  }
+);
 
 export default Alimento;
